Feat: API통신 모듈화, 인터셉터 적용, request에 토큰 설정
Former-commit-id: 92c7c9ed5919775c492681d6b5e582c37e77a591

+++ backend/src/main/java/com/moi/anitime/model/entity/animal/Animal.java
@@ -16,7 +16,7 @@ public class Animal {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "desertionno")
-    int desertionNo;
+    long desertionNo;
     @Column(name = "shelterno")
     int shelterNo;
